aethernav/requirements.txt
numpy>=1.20.0
rclpy>=1.0.0
launch>=1.0.0
launch_ros>=0.19.0
geometry_msgs>=2.0.0
nav_msgs>=2.0.0
pytest>=6.0.0
pytest-ros>=0.3.0
setuptools>=58.0.0
psutil>=5.8.0
matplotlib>=3.4.0  # For visualization
pyyaml>=5.4.0  # For configuration files